# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Tech-Admin-Api",
        "contact": {
            "name": "Tech-Admin",
            "url": "  http://tech-admin.com",
            "email": "  @gmail.com"
        },
        "version": "0.1"
    },
    "paths": {
...
        "/api/acesso/{id}": {
            "get": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Mostra um acesso",
                "description": "Mostra um acesso",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o acesso"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "put": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Atualiza um acesso",
                "description": "Atualiza um acesso",
                "operationId": "update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome do acesso",
                                        "type": "string"
                                    },
                                    "descricao": {
                                        "description": "Descrição do acesso",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status do acesso",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o acesso atualizado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Acesso"
                ],
                "summary": "Deleta um acesso",
                "description": "Deleta um acesso",
                "operationId": "destroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do acesso",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o acesso deletado"
                    },
                    "default": {
                        "description": "Caso não encontre o acesso"
                    }
                }
            }
        },
...
        "/api/produto": {
            "get": {
                "tags": [
                    "Produto"
                ],
                "summary": "Lista todos os produtos",
                "description": "Retorna todos os produtos",
                "operationId": "indexProduto",
                "responses": {
                    "200": {
                        "description": "Retorna todos os produtos"
                    },
                    "default": {
                        "description": "Caso não encontre o produto"
                    }
                }
            },
            "post": {
                "tags": [
                    "Produto"
                ],
                "summary": "Cria um novo produto",
                "description": "Cria um novo produto",
                "operationId": "storeProduto",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome do produto",
                                        "type": "string",
                                        "example": "Produto Teste"
                                    },
                                    "descricao": {
                                        "description": "Descrição do produto",
                                        "type": "string",
                                        "example": "Produto Teste"
                                    },
                                    "valor": {
                                        "description": "Valor do produto",
                                        "type": "number",
                                        "example": "10.00"
                                    },
                                    "quantidade": {
                                        "description": "Quantidade do produto",
                                        "type": "number",
                                        "example": "10"
                                    },
                                    "id_estoque": {
                                        "description": "ID do estoque",
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o produto criado"
                    },
                    "default": {
                        "description": "Caso não encontre o produto"
                    }
                }
            }
        },
        "/api/produto/{id}": {
            "get": {
                "tags": [
                    "Produto"
                ],
                "summary": "Mostra um produto",
                "description": "Mostra um produto",
                "operationId": "showProduto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o produto"
                    },
                    "default": {
                        "description": "Caso não encontre o produto"
                    }
                }
            },
            "put": {
                "tags": [
                    "Produto"
                ],
                "summary": "Atualiza um produto",
                "description": "Atualiza um produto",
                "operationId": "updateProduto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nome": {
                                        "description": "Nome do produto",
                                        "type": "string",
                                        "example": "Produto Teste"
                                    },
                                    "descricao": {
                                        "description": "Descrição do produto",
                                        "type": "string",
                                        "example": "Produto Teste"
                                    },
                                    "valor": {
                                        "description": "Valor do produto",
                                        "type": "number",
                                        "example": "10.00"
                                    },
                                    "quantidade": {
                                        "description": "Quantidade do produto",
                                        "type": "number",
                                        "example": "10"
                                    },
                                    "id_estoque": {
                                        "description": "ID do estoque",
                                        "type": "number",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retorna o produto atualizado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Produto"
                ],
                "summary": "Remove um produto",
                "description": "Remove um produto",
                "operationId": "deleteProduto",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do produto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retorna o produto removido"
                    }
                }
            }
        },
...
    }
}